fish: xv - Select by resolution instead of size
Selecting the preferred format by filesize isn't ideal, long videos may
still exceed the size at long run-times but aren't too heavy for my
hardware, and short videos might be under a size limit while having a
resolution that is hard for my hardware to cope with.

Now that youtube-dl is unmaintained it's time to say goodbye to having
a separate function for yt-dlp.